refactor: Move useButtonStates to lib/hooks folder and update imports. Remove unneeded imports, comments.
