+++ app/lib/types.ts
@@ -286,10 +286,6 @@ export interface EmailData {
   text?: string;
 }
 
-/**
- * Represents the state of chord progressions
- * Maps position indices to chord symbols
- */
 export interface ProgressionState {
   [key: string]: string;
 }
