+++ firebase/authContext.tsx
@@ -2,7 +2,7 @@
 import { Stack } from "@mui/material";
 import Box from "@mui/material/Box";
 import { User, onAuthStateChanged } from "firebase/auth";
-import { ReactNode, useContext, useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import { auth } from "../firebase/config";
 import CreateAuthContext from "./createAuthContext";
 import { AuthContextType } from "@/app/lib/types";
